refactor(routes): move Velociraptor event monitoring to its own module
routes/velociraptor.ts was frozen at 888 lines. The seam is real rather than
convenient: everything else in that module is REQUEST/RESPONSE — run a query,
launch a hunt, collect from a host, import results — a bounded exchange that
finishes. A monitor is a standing subscription with a lifecycle of its own: it
is created, polled on a timer, stopped, restarted and removed, and it owns
background timers (scheduleVeloMonitor / stopVeloMonitorTimer) nothing else
there touches.

139 lines moved verbatim — route bodies, status codes and messages unchanged,
only the registration point, following the register*Routes convention already
used across routes/.

The `?refresh=1` predicate is duplicated rather than exported: neither module
should own a two-line query-string check on the other's behalf, and exporting it
would have made the artifact catalog and the monitor catalog share a dependency
they otherwise do not have.

routes/velociraptor.ts: 888 -> 751, under the 800-line threshold, so it leaves
the ledger entirely. 12 -> 11 files.

Worth recording: running Prettier over this file inflated it 753 -> 1127, which
would have blown the very cap this change exists to clear. It is one of the
files that predates the formatter, and reformatting it is a separate decision
from splitting it — bundling the two would bury a 139-line move in a 1,100-line
diff. Same trap as the earlier bounded-output fix.

Verified: typecheck, eslint, prettier clean; routes suites 962 passed; boundary,
import-cycle and a11y ratchets unmoved.
